refactoring into seperate components
